Move hive monitor checks into their own document
