+++ pkg/monitor/hive/clustersync.go
@@ -1,4 +1,4 @@
-package cluster
+package hive
 
 // Copyright (c) Microsoft Corporation.
 // Licensed under the Apache License 2.0.
