+++ pkg/monitor/hive/hiveregistrationstatus.go
@@ -1,4 +1,4 @@
-package cluster
+package hive
 
 // Copyright (c) Microsoft Corporation.
 // Licensed under the Apache License 2.0.
@@ -9,11 +9,7 @@ import (
 
 	corev1 "k8s.io/api/core/v1"
 
-	"sigs.k8s.io/controller-runtime/pkg/client"
-
 	hivev1 "github.com/openshift/hive/apis/hive/v1"
-
-	"github.com/Azure/ARO-RP/pkg/hive"
 )
 
 var clusterDeploymentConditionsExpected = map[hivev1.ClusterDeploymentConditionType]corev1.ConditionStatus{
@@ -32,7 +28,7 @@ func (mon *Monitor) emitHiveRegistrationStatus(ctx context.Context) error {
 		return fmt.Errorf("cluster %s not adopted. No namespace in the clusterdocument", mon.oc.Name)
 	}
 
-	cd, err := mon.retrieveClusterDeployment(ctx)
+	cd, err := mon.hiveClusterManager.GetClusterDeployment(ctx, mon.oc)
 	if err != nil {
 		return err
 	}
@@ -40,19 +36,6 @@ func (mon *Monitor) emitHiveRegistrationStatus(ctx context.Context) error {
 	conditions := mon.filterClusterDeploymentConditions(ctx, cd, clusterDeploymentConditionsExpected)
 	mon.emitFilteredClusterDeploymentMetrics(conditions)
 	return nil
-}
-
-func (mon *Monitor) retrieveClusterDeployment(ctx context.Context) (*hivev1.ClusterDeployment, error) {
-	cd := &hivev1.ClusterDeployment{}
-	err := mon.hiveclientset.Get(ctx, client.ObjectKey{
-		Namespace: mon.oc.Properties.HiveProfile.Namespace,
-		Name:      hive.ClusterDeploymentName,
-	}, cd)
-	if err != nil {
-		return nil, err
-	}
-
-	return cd, nil
 }
 
 func (mon *Monitor) filterClusterDeploymentConditions(ctx context.Context, cd *hivev1.ClusterDeployment, clusterDeploymentConditionsExpected map[hivev1.ClusterDeploymentConditionType]corev1.ConditionStatus) []hivev1.ClusterDeploymentCondition {
